LPS-150106 Move TranslateActionBar components into their folder
