+++ modules/apps/translation/translation-web/src/main/resources/META-INF/resources/js/translate/components/TranslateActionBar/TranslateLanguagesSelector.js
